+++ UnitTest/Audio/DriveAudioMixerTests.cpp
@@ -1,7 +1,4 @@
-#include "../CassoEmuCore/Pch.h"
-
-#include <CppUnitTest.h>
-
+#include "Pch.h"
 #include "Audio/DriveAudioMixer.h"
 #include "Audio/IDriveAudioSource.h"