+++ UnitTest/CpuIrqTests.cpp
@@ -1,6 +1,5 @@
 #include "Pch.h"
 
-#include "CppUnitTest.h"
 #include "Cpu6502.h"
 #include "ICpu.h"
 #include "TestHelpers.h"