+++ UnitTest/EmuTests/PrngTests.cpp
@@ -1,7 +1,4 @@
-#include "../CassoEmuCore/Pch.h"
-
-#include <CppUnitTest.h>
-
+#include "Pch.h"
 #include "Core/Prng.h"
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;